Refactoring code/tests to create a package
